Created a collection of monitor database. Retrieved the datas and displayed them on the table created by tabulator

+++ monitor/models.py
@@ -1,65 +1,65 @@
 from djongo import models
+from decimal import Decimal
 
-
-class ExchangeRate(models.Model):  # usd/yen,canad/yen, gen/yen, euro/yen, ind/yen
+class ExchangeRate(models.Model):
     timestamp = models.DateTimeField()
     currency_pair = models.CharField(max_length=255)
-    price = models.FloatField()
-    change_price = models.FloatField()
-    change_rate = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
+    change_price = models.DecimalField(max_digits=10, decimal_places=2)
+    change_rate = models.DecimalField(max_digits=10, decimal_places=2)
     updated_at = models.DateTimeField()
 
 
-class StockPrice(models.Model):  # nikkeiaverage, topix, nydau, nasdaq, s&p500
+class StockPrice(models.Model):
     timestamp = models.DateTimeField()
     index_name = models.CharField(max_length=255)
-    price = models.FloatField()
-    change_price = models.FloatField()
-    change_rate = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
+    change_price = models.DecimalField(max_digits=10, decimal_places=2)
+    change_rate = models.DecimalField(max_digits=10, decimal_places=2)
     updated_at = models.DateTimeField()
 
 
-class CommodityPrice(models.Model):  # wticrude, gold, platina, silver
+class CommodityPrice(models.Model):
     timestamp = models.DateTimeField()
     commodity_name = models.CharField(max_length=255)
-    price = models.FloatField()
-    change_price = models.FloatField()
-    change_rate = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
+    change_price = models.DecimalField(max_digits=10, decimal_places=2)
+    change_rate = models.DecimalField(max_digits=10, decimal_places=2)
     updated_at = models.DateTimeField()
 
-class BondRate(models.Model):  # 国債
+class BondRate(models.Model):
     timestamp = models.DateTimeField()
     bond_name = models.CharField(max_length=255)
-    annual_interest_rate = models.FloatField()
-    change_interest_rate = models.FloatField()
-    change_interest_rateByRate = models.FloatField()
+    annual_interest_rate = models.DecimalField(max_digits=10, decimal_places=2)
+    change_interest_rate = models.DecimalField(max_digits=10, decimal_places=2)
+    change_interest_rateByRate = models.DecimalField(max_digits=10, decimal_places=2)
     updated_at = models.DateTimeField()
 
 
 class FedRate(models.Model):
     timestamp = models.DateTimeField()
     fed_name = models.CharField(max_length=255)
-    fedRate = models.FloatField()
-    change_fedRate = models.FloatField()
-    change_fedRateByRate = models.FloatField()
+    fedRate = models.DecimalField(max_digits=10, decimal_places=2)
+    change_fedRate = models.DecimalField(max_digits=10, decimal_places=2)
+    change_fedRateByRate = models.DecimalField(max_digits=10, decimal_places=2)
     updated_at = models.DateTimeField()
 
 
-class ShippingIndex(models.Model):  # バルチック、定期船運賃
+class ShippingIndex(models.Model):
     timestamp = models.DateTimeField()
     index_name = models.CharField(max_length=255)
-    index_value = models.FloatField()
-    change_price = models.FloatField()
-    change_rate = models.FloatField()
+    index_value = models.DecimalField(max_digits=10, decimal_places=2)
+    change_price = models.DecimalField(max_digits=10, decimal_places=2)
+    change_rate = models.DecimalField(max_digits=10, decimal_places=2)
     updated_at = models.DateTimeField()
 
 
-class GdpData(models.Model):  # GDP
+class GdpData(models.Model):
     timestamp = models.DateTimeField()
     country_name = models.CharField(max_length=255)
-    gdp = models.FloatField()
-    change_gdp = models.FloatField()
-    change_rate = models.FloatField()
+    gdp = models.DecimalField(max_digits=10, decimal_places=2)
+    change_gdp = models.DecimalField(max_digits=10, decimal_places=2)
+    change_rate = models.DecimalField(max_digits=10, decimal_places=2)
     updated_at = models.DateTimeField()
